scripts/requirements-ingest.txt
supabase>=2.3.0
openai>=1.12.0
aiohttp>=3.9.3
python-dotenv>=1.0.1
google-generativeai>=0.8.2
pypdf2
